+++ core/src/main/java/es/urjc/etsii/grafo/solver/improve/LocalSearch.java
@@ -3,7 +3,7 @@ package es.urjc.etsii.grafo.solver.improve;
 import es.urjc.etsii.grafo.io.Instance;
 import es.urjc.etsii.grafo.solution.Move;
 import es.urjc.etsii.grafo.solution.MoveComparator;
-import es.urjc.etsii.grafo.solution.Neighborhood;
+import es.urjc.etsii.grafo.solution.neighborhood.Neighborhood;
 import es.urjc.etsii.grafo.solution.Solution;
 
 import java.util.Arrays;