+++ core/src/main/java/es/urjc/etsii/grafo/solver/irace/runners/RLangRunner.java
@@ -1,0 +1,5 @@
+package es.urjc.etsii.grafo.solver.irace.runners;
+
+public interface RLangRunner {
+    void execute(String rCode);
+}