+++ core/src/main/java/es/urjc/etsii/grafo/solver/services/events/EventPublisher.java
@@ -14,6 +14,12 @@ public class EventPublisher {
     private static Logger log = Logger.getLogger(EventPublisher.class.getName());
     private static ApplicationEventPublisher publisher;
 
+    /**
+     * Disable event propagation when executing in irace mode
+     */
+
+    private static boolean blockEvents = false;
+
     protected EventPublisher(ApplicationEventPublisher publisher) {
         EventPublisher.publisher = publisher;
     }
@@ -23,7 +29,18 @@ public class EventPublisher {
      * @param event Event to propagate
      */
     public static void publishEvent(MorkEvent event){
+        if(blockEvents){
+            log.fine("Event system disabled: "+event);
+        }
         log.fine("Publishing event: " + event);
         publisher.publishEvent(event);
+    }
+
+    public static void block(){
+        EventPublisher.blockEvents = true;
+    }
+
+    public static void unblock(){
+        EventPublisher.blockEvents = false;
     }
 }